backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt[crypto]==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlmodel==0.0.31
asyncpg==0.31.0
alembic==1.17.2
psycopg2-binary==2.9.11
